requirements.txt
## System dependencies and utilities
clinner
health_check
ipython
aiohttp

## Django
Django
django-configurations
django-extensions

## Frontend
whitenoise

## Django REST Framework
djangorestframework
